requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gradio==4.9.0
transformers==4.35.0
datasets==2.14.5
torch==2.1.0
soundfile==0.12.1
python-dotenv==1.0.0
google-generativeai==0.3.1
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.4.2
librosa==0.10.1 